Address PR review comments (round 10)
- Remove dead --yes/-y flag from import_file (never read or forwarded)
- Add 'failed' status in finalize_import_batch when rows_imported=0
- Fix docstring: "float" → "Decimal" in _generate_transaction_ids
- Fix test: rename to test_finalize_sets_partial_status, assert concrete param
- Add test_finalize_sets_complete_status and test_finalize_sets_failed_status
- Fix mock setup in test_revert_deletes_rows: use side_effect for COUNT query

+++ src/moneybin/extractors/tabular/transforms.py
@@ -515,7 +515,7 @@ def _generate_transaction_ids(
         src_txn_id_col: Column name holding source transaction IDs, or None.
         account_id: Account identifier.
         original_date_strs: Raw date strings, one per row.
-        parsed_amounts: Parsed float amounts (None for invalid rows).
+        parsed_amounts: Parsed Decimal amounts (None for invalid rows).
         descriptions: Description strings, one per row.
         row_numbers: 1-based row numbers.
         source_type: File type prefix for content-hash IDs.
